+++ Tests/OCTRealmTests.h
@@ -24,5 +24,6 @@
 @property (strong, nonatomic) OCTRealmManager *realmManager;
 
 - (OCTFriend *)createFriend;
+- (OCTFriend *)createFriendWithFriendNumber:(OCTToxFriendNumber)friendNumber;
 
 @end